completed ft_printf

+++ ft_put_ptr.c
@@ -6,14 +6,14 @@
 /*   By: dravi-ch <dravi-ch@student.42kl.edu.my>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/06/02 17:29:30 by dravi-ch          #+#    #+#             */
-/*   Updated: 2023/06/03 14:23:22 by dravi-ch         ###   ########.fr       */
+/*   Updated: 2023/06/05 14:51:57 by dravi-ch         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
 /*Writes pointer address and returns the lenght of the address*/
-int	ft_put_ptr(unsigned long long ptr)
+int	ft_put_ptr(uintptr_t ptr)
 {
 	int	len;
